feat: extract TrustyAI types into dedicated shared package
Move TrustyInstallState, BiasMetricType, BaseMetric*, and
BiasMetricConfig from model-serving/shared/types into the new
packages/trustyai/ workspace. This allows future consumers to
depend on TrustyAI domain types without coupling to model-serving.

Prometheus metrics types remain in model-serving/shared/types.
