+++ frontend/src/api/trustyai/rawTypes.ts
@@ -1,4 +1,4 @@
-import { type BaseMetric, BiasMetricType } from '@odh-dashboard/model-serving/shared/types';
+import { type BaseMetric, BiasMetricType } from '@odh-dashboard/trustyai/types';
 
 export enum DataTypes {
   BOOL = 'BOOL',
