+++ frontend/src/concepts/trustyai/utils.ts
@@ -1,6 +1,6 @@
 import { getConditionForType } from '@odh-dashboard/k8s-core';
 import { FetchState } from '@odh-dashboard/ui-core/hooks/useFetchState';
-import { BiasMetricConfig, TrustyInstallState } from '@odh-dashboard/model-serving/shared/types';
+import { BiasMetricConfig, TrustyInstallState } from '@odh-dashboard/trustyai/types';
 import { BaseMetricListResponse } from '#~/api';
 import { TrustyStatusStates } from '#~/concepts/trustyai/types';
 import { TrustyAIKind } from '#~/k8sTypes';
