+++ frontend/src/pages/projects/projectSettings/__tests__/ModelBiasSettingsCard.spec.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
-import { TrustyInstallState } from '@odh-dashboard/model-serving/shared/types';
+import { TrustyInstallState } from '@odh-dashboard/trustyai/types';
 import { useTrustySettingsAccessAllowed } from '#~/concepts/trustyai/content/useTrustySettingsAccessAllowed';
 import useManageTrustyAICR from '#~/concepts/trustyai/useManageTrustyAICR';
 import ModelBiasSettingsCard from '#~/pages/projects/projectSettings/ModelBiasSettingsCard';
