+++ packages/model-serving/package.json
@@ -75,6 +75,7 @@
     "@odh-dashboard/k8s-core": "*",
     "@odh-dashboard/model-registry": "*",
     "@odh-dashboard/plugin-core": "*",
+    "@odh-dashboard/trustyai": "*",
     "@odh-dashboard/ui-core": "*",
     "lodash-es": "^4.18.1",
     "yaml": "^2.9.0",
