+++ packages/model-serving/src/components/metrics/bias/BiasConfigurationPage/BiasConfigurationModal/ManageBiasConfigurationModal.tsx
@@ -11,7 +11,7 @@ import {
 import DashboardHelpTooltip from '@odh-dashboard/ui-core/components/DashboardHelpTooltip';
 import { TrustyAIContext } from '@odh-dashboard/internal/concepts/trustyai/context/TrustyAIContext';
 import DashboardModalFooter from '@odh-dashboard/ui-core/components/DashboardModalFooter';
-import { BiasMetricConfig, BiasMetricType } from '@odh-dashboard/model-serving/shared/types';
+import { BiasMetricConfig, BiasMetricType } from '@odh-dashboard/trustyai/types';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import useBiasConfigurationObject from './useBiasConfigurationObject';
 import MetricTypeField from './MetricTypeField';
