+++ packages/model-serving/src/components/metrics/bias/BiasConfigurationPage/BiasConfigurationModal/MetricTypeField.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { FormGroup } from '@patternfly/react-core';
 import { asEnumMember, enumIterator } from '@odh-dashboard/foundation';
 import SimpleSelect, { SimpleSelectOption } from '@odh-dashboard/ui-core/components/SimpleSelect';
-import { BiasMetricType } from '@odh-dashboard/model-serving/shared/types';
+import { BiasMetricType } from '@odh-dashboard/trustyai/types';
 import { METRIC_TYPE_DESCRIPTION, METRIC_TYPE_DISPLAY_NAME } from '../../../const';
 import { isMetricType } from '../../../utils';
