+++ packages/model-serving/src/components/metrics/bias/BiasTab.tsx
@@ -15,7 +15,7 @@ import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import MetricsPageToolbar from '@odh-dashboard/internal/concepts/metrics/MetricsPageToolbar';
 import { useModelBiasData } from '@odh-dashboard/internal/concepts/trustyai/context/useModelBiasData';
 import DashboardExpandableSection from '@odh-dashboard/ui-core/components/DashboardExpandableSection';
-import { TrustyInstallState } from '@odh-dashboard/model-serving/shared/types';
+import { TrustyInstallState } from '@odh-dashboard/trustyai/types';
 import BiasMetricConfigSelector from './BiasMetricConfigSelector';
 import BiasChart from './BiasChart';
 import EmptyBiasConfigurationCard from './BiasConfigurationPage/EmptyBiasConfigurationCard';
