+++ packages/model-serving/src/components/metrics/types.ts
@@ -1,11 +1,8 @@
 import { DomainTuple, ForAxes } from 'victory-core';
 import { ChartThemeDefinitionInterface } from '@patternfly/react-charts/victory';
 import { FetchStateObject } from '@odh-dashboard/ui-core/hooks/useFetch';
-import {
-  PrometheusQueryRangeResultValue,
-  BiasMetricType,
-  BiasMetricConfig,
-} from '@odh-dashboard/model-serving/shared/types';
+import { PrometheusQueryRangeResultValue } from '@odh-dashboard/model-serving/shared/types';
+import { BiasMetricType, BiasMetricConfig } from '@odh-dashboard/trustyai/types';
 import { ModelMetricType } from './ModelServingMetricsContext';
 
 export type TranslatePoint = (line: GraphMetricPoint) => GraphMetricPoint;
