+++ packages/model-serving/src/components/metrics/utils.tsx
@@ -6,13 +6,13 @@ import { Link } from 'react-router-dom';
 import { isEnumMember } from '@odh-dashboard/foundation';
 import { TimeframeTitle, BreadcrumbItemType } from '@odh-dashboard/ui-core/types/metrics';
 import { PROMETHEUS_REQUEST_RESOLUTION } from '@odh-dashboard/ui-core/utilities/metrics';
+import { PrometheusQueryRangeResultValue } from '@odh-dashboard/model-serving/shared/types';
 import {
-  PrometheusQueryRangeResultValue,
   BaseMetricRequest,
   BaseMetricRequestInput,
   BiasMetricType,
   BiasMetricConfig,
-} from '@odh-dashboard/model-serving/shared/types';
+} from '@odh-dashboard/trustyai/types';
 import type { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
 import { BIAS_CHART_CONFIGS, BIAS_THRESHOLD_COLOR } from './const';
 import {
